+++ packages/manager/src/store/globalErrors/types.ts
@@ -2,4 +2,4 @@ interface S {
   account_unactivated: boolean;
 }
 
-export type State = Partial<S>
+export type State = Partial<S>;